data_analysis/requirements.txt
pandas>=2.2.0
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0
dash>=2.16.0
Pillow>=10.2.0
numpy>=1.26.0
tqdm>=4.66.0
